requirements.txt
dspy-ai[chromadb]
pydantic
sqlmodel
fastapi
uvicorn[standard]